6335 | 91 files | M DumpSource2/schemas/client/CDOTA_Ability_DarkSeer_WallOfReplica.h,  M DumpSource2/schemas/client/CDOTA_Ability_Elder_Titan_EarthSplitter.h,  M DumpSource2/schemas/client/CDOTA_Ability_Magnataur_Skewer.h,  M DumpSource2/schemas/client/CDOTA_Ability_Marci_Unleash.h,  M DumpSource2/schemas/client/CDOTA_Ability_VoidSpirit_AetherRemnant.h,  M DumpSource2/schemas/client/CDOTA_Modifier_Beastmaster_CallOfTheWild_Hawk.h,  M DumpSource2/schemas/client/CDOTA_Modifier_DeathProphet_Exorcism.h,  M DumpSource2/schemas/client/CDOTA_Modifier_Enchantress_NaturesAttendants.h,  M DumpSource2/schemas/client/CDOTA_Modifier_Enigma_BlackHoleThinker.h,  M DumpSource2/schemas/client/CDOTA_Modifier_Enigma_BlackHole_Pull.h,  M DumpSource2/schemas/client/CDOTA_Modifier_Hoodwink_Sharpshooter_Windup.h,  M DumpSource2/schemas/client/CDOTA_Modifier_Jakiro_LiquidFire.h,  M DumpSource2/schemas/client/CDOTA_Modifier_Jakiro_Macropyre_Thinker.h,  M DumpSource2/schemas/client/CDOTA_Modifier_Kez_FalconRush.h,  M DumpSource2/schemas/client/CDOTA_M
https://steamdb.info/patchnotes/17508113/

+++ DumpSource2/schemas/client/CDOTA_Ability_DarkSeer_WallOfReplica.h
@@ -1,4 +1,4 @@
 class CDOTA_Ability_DarkSeer_WallOfReplica
 {
-	int32 width;
+	float32 width;
 };
