+++ DumpSource2/schemas/client/CDOTA_Ability_Elder_Titan_EarthSplitter.h
@@ -1,8 +1,8 @@
 class CDOTA_Ability_Elder_Titan_EarthSplitter
 {
-	int32 crack_width;
-	int32 crack_distance;
-	int32 speed;
-	int32 vision_width;
+	float32 crack_width;
+	float32 crack_distance;
+	float32 speed;
+	float32 vision_width;
 	float32 crack_time;
 };
