+++ DumpSource2/schemas/client/CDOTA_Ability_Magnataur_Skewer.h
@@ -1,10 +1,10 @@
 class CDOTA_Ability_Magnataur_Skewer
 {
 	float32 skewer_radius;
-	int32 skewer_speed;
-	int32 tree_radius;
-	int32 tree_hit_damage;
-	int32 cliff_hit_damage;
+	float32 skewer_speed;
+	float32 tree_radius;
+	float32 tree_hit_damage;
+	float32 cliff_hit_damage;
 	int32 terrain_hit_increase_pct;
 	float32 terrain_hit_cooldown;
 	int32 m_nTargetsHit;
