+++ DumpSource2/schemas/client/CDOTA_Ability_VoidSpirit_AetherRemnant.h
@@ -1,7 +1,7 @@
 class CDOTA_Ability_VoidSpirit_AetherRemnant
 {
-	int32 start_radius;
-	int32 end_radius;
-	int32 remnant_watch_distance;
-	int32 projectile_speed;
+	float32 start_radius;
+	float32 end_radius;
+	float32 remnant_watch_distance;
+	float32 projectile_speed;
 };
