+++ DumpSource2/schemas/client/CDOTA_Modifier_Enchantress_NaturesAttendants.h
@@ -2,8 +2,8 @@ class CDOTA_Modifier_Enchantress_NaturesAttendants
 {
 	float32 heal_interval;
 	float32 heal;
-	int32 radius;
-	int32 movespeed;
+	float32 radius;
+	float32 movespeed;
 	int32 wisp_count;
 	ParticleIndex_t m_nWispFXIndex;
 	bool m_bAutoWisps;
