+++ DumpSource2/schemas/client/CDOTA_Modifier_Enigma_BlackHoleThinker.h
@@ -1,5 +1,5 @@
 class CDOTA_Modifier_Enigma_BlackHoleThinker
 {
-	int32 radius;
+	float32 radius;
 	ParticleIndex_t n_FXIndex;
 };
