+++ DumpSource2/schemas/client/CDOTA_Modifier_Enigma_BlackHole_Pull.h
@@ -1,8 +1,8 @@
 class CDOTA_Modifier_Enigma_BlackHole_Pull
 {
 	int32 pull_speed;
-	int32 radius;
-	int32 damage;
+	float32 radius;
+	float32 damage;
 	float32 tick_rate;
 	float32 pull_rotate_speed;
 	float32 animation_rate;
