+++ DumpSource2/schemas/client/CDOTA_Modifier_Hoodwink_Sharpshooter_Windup.h
@@ -1,6 +1,6 @@
 class CDOTA_Modifier_Hoodwink_Sharpshooter_Windup
 {
-	int32 arrow_vision;
+	float32 arrow_vision;
 	float32 max_charge_time;
 	float32 turn_rate;
 	float32 base_power;
