+++ DumpSource2/schemas/client/CDOTA_Modifier_Kez_FalconRush.h
@@ -1,8 +1,8 @@
 class CDOTA_Modifier_Kez_FalconRush
 {
 	int32 buff_evasion_pct;
-	int32 rush_speed;
-	int32 rush_range;
+	float32 rush_speed;
+	float32 rush_range;
 	bool m_bInRush;
 	ParticleIndex_t m_nFXIndex;
 	CHandle< C_BaseEntity > m_hTarget;
