+++ DumpSource2/schemas/client/CDOTA_Modifier_Mirana_Starfall_Thinker.h
@@ -1,5 +1,5 @@
 class CDOTA_Modifier_Mirana_Starfall_Thinker
 {
-	int32 starfall_secondary_radius;
+	float32 starfall_secondary_radius;
 	float32 starstruck_duration;
 };
