+++ DumpSource2/schemas/client/CDOTA_Modifier_Shredder_Flamethrower.h
@@ -1,8 +1,8 @@
 class CDOTA_Modifier_Shredder_Flamethrower
 {
-	int32 length;
-	int32 width;
-	int32 damage_per_second;
+	float32 length;
+	float32 width;
+	float32 damage_per_second;
 	ParticleIndex_t m_nBeamFXIndex;
 	CHandle< C_BaseEntity > m_hBeamEnd;
 	GameTime_t m_flLastHit;
