+++ DumpSource2/schemas/client/CDOTA_Modifier_Spectre_Dispersion.h
@@ -1,8 +1,8 @@
 class CDOTA_Modifier_Spectre_Dispersion
 {
 	int32 damage_reflection_pct;
-	int32 min_radius;
-	int32 max_radius;
+	float32 min_radius;
+	float32 max_radius;
 	int32 activation_bonus_pct;
 	GameTime_t m_flLastDispersionPulseEffectTime;
 	float32 m_fDamageCounter;
