+++ DumpSource2/schemas/client/CDOTA_Modifier_Spectre_SpectralDagger.h
@@ -1,6 +1,6 @@
 class CDOTA_Modifier_Spectre_SpectralDagger
 {
-	int32 path_radius;
+	float32 path_radius;
 	float32 hero_path_duration;
 	Vector m_vLastPosition;
 };
