+++ DumpSource2/schemas/client/CDOTA_Modifier_Viper_CorrosiveSkin.h
@@ -1,6 +1,6 @@
 class CDOTA_Modifier_Viper_CorrosiveSkin
 {
 	int32 bonus_magic_resistance;
-	int32 max_range;
+	float32 max_range;
 	int32 nethertoxin_bonus_pct;
 };
