+++ DumpSource2/schemas/client/CDOTA_Modifier_Warlock_Rain_Of_Chaos_Thinker.h
@@ -2,6 +2,6 @@ class CDOTA_Modifier_Warlock_Rain_Of_Chaos_Thinker
 {
 	float32 stun_duration;
 	bool bHasScepter;
-	int32 aoe;
+	float32 aoe;
 	int32 m_iCount;
 };
