+++ DumpSource2/schemas/client/C_DOTA_Ability_Earthshaker_EchoSlam.h
@@ -1,8 +1,8 @@
 class C_DOTA_Ability_Earthshaker_EchoSlam
 {
 	float32 echo_slam_damage_range;
-	int32 echo_slam_echo_search_range;
-	int32 echo_slam_echo_range;
-	int32 echo_slam_echo_damage;
-	int32 echo_slam_initial_damage;
+	float32 echo_slam_echo_search_range;
+	float32 echo_slam_echo_range;
+	float32 echo_slam_echo_damage;
+	float32 echo_slam_initial_damage;
 };
