+++ DumpSource2/schemas/client/C_DOTA_Ability_Lich_Ice_Spire.h
@@ -1,5 +1,5 @@
 class C_DOTA_Ability_Lich_Ice_Spire
 {
-	int32 aura_radius;
+	float32 aura_radius;
 	float32 duration;
 };
