+++ DumpSource2/schemas/client/C_DOTA_Ability_Mirana_Arrow.h
@@ -1,6 +1,6 @@
 class C_DOTA_Ability_Mirana_Arrow
 {
-	int32 scepter_radius;
+	float32 scepter_radius;
 	Vector m_vStartPos;
 	ParticleIndex_t m_nFXIndex;
 	CUtlVector< CHandle< C_BaseEntity > >[2] hAlreadyHitList;
