+++ DumpSource2/schemas/client/C_DOTA_Ability_Warlock_RainOfChaos.h
@@ -1,5 +1,5 @@
 class C_DOTA_Ability_Warlock_RainOfChaos
 {
-	int32 aoe;
+	float32 aoe;
 	bool m_bHasScepter;
 };
