+++ DumpSource2/schemas/server/CDOTA_Ability_DarkSeer_WallOfReplica.h
@@ -1,5 +1,5 @@
 class CDOTA_Ability_DarkSeer_WallOfReplica
 {
-	int32 width;
+	float32 width;
 	Vector m_vEndpoint;
 };
