+++ DumpSource2/schemas/server/CDOTA_Ability_EarthSpirit_BoulderSmash.h
@@ -1,10 +1,10 @@
 class CDOTA_Ability_EarthSpirit_BoulderSmash
 {
-	int32 speed;
-	int32 rock_damage;
+	float32 speed;
+	float32 rock_damage;
 	float32 creep_multiplier;
-	int32 radius;
-	int32 rock_search_aoe;
+	float32 radius;
+	float32 rock_search_aoe;
 	float32 unit_distance;
 	float32 rock_distance;
 	int32 m_nProjectileID;
