+++ DumpSource2/schemas/server/CDOTA_Ability_Invoker_ChaosMeteor.h
@@ -1,8 +1,8 @@
 class CDOTA_Ability_Invoker_ChaosMeteor
 {
-	int32 area_of_effect;
+	float32 area_of_effect;
 	float32 damage_interval;
-	int32 vision_distance;
+	float32 vision_distance;
 	float32 end_vision_duration;
 	float32 main_damage;
 	float32 burn_duration;
