+++ DumpSource2/schemas/server/CDOTA_Ability_Lich_Ice_Spire.h
@@ -1,5 +1,5 @@
 class CDOTA_Ability_Lich_Ice_Spire
 {
-	int32 aura_radius;
+	float32 aura_radius;
 	float32 duration;
 };
