+++ DumpSource2/schemas/server/CDOTA_Ability_Mirana_Arrow.h
@@ -1,7 +1,7 @@
 class CDOTA_Ability_Mirana_Arrow
 {
 	CHandle< CDOTABaseAbility > m_hSourceAbility;
-	int32 scepter_radius;
+	float32 scepter_radius;
 	Vector m_vStartPos;
 	ParticleIndex_t m_nFXIndex;
 	CUtlVector< CHandle< CBaseEntity > >[2] hAlreadyHitList;
