+++ DumpSource2/schemas/server/CDOTA_Ability_MonkeyKing_Boundless_Strike.h
@@ -1,7 +1,7 @@
 class CDOTA_Ability_MonkeyKing_Boundless_Strike
 {
-	int32 strike_cast_range;
-	int32 strike_radius;
+	float32 strike_cast_range;
+	float32 strike_radius;
 	int32 spring_channel_pct;
 	ParticleIndex_t m_nFXIndex;
 	bool m_bIsAltCastState;
