+++ DumpSource2/schemas/server/CDOTA_Ability_Phoenix_LaunchFireSpirit.h
@@ -1,7 +1,7 @@
 class CDOTA_Ability_Phoenix_LaunchFireSpirit
 {
-	int32 spirit_speed;
+	float32 spirit_speed;
 	float32 duration;
-	int32 radius;
+	float32 radius;
 	ParticleIndex_t m_nFXIndex;
 };
