+++ DumpSource2/schemas/server/CDOTA_Ability_VengefulSpirit_WaveOfTerror.h
@@ -1,12 +1,12 @@
 class CDOTA_Ability_VengefulSpirit_WaveOfTerror
 {
-	int32 wave_width;
+	float32 wave_width;
 	float32 wave_speed;
 	int32 m_iProjectile;
 	float32 vision_aoe;
 	float32 vision_duration;
 	int32 steal_pct;
-	int32 damage;
+	float32 damage;
 	int32 m_nNumHeroesHit;
 	CountdownTimer m_ViewerTimer;
 };
