+++ DumpSource2/schemas/server/CDOTA_Ability_Warlock_RainOfChaos.h
@@ -1,5 +1,5 @@
 class CDOTA_Ability_Warlock_RainOfChaos
 {
-	int32 aoe;
+	float32 aoe;
 	bool m_bHasScepter;
 };
