+++ DumpSource2/schemas/server/CDOTA_Modifier_Jakiro_Macropyre_Thinker.h
@@ -2,8 +2,8 @@ class CDOTA_Modifier_Jakiro_Macropyre_Thinker
 {
 	Vector m_vPathDir;
 	float32 burn_interval;
-	int32 path_radius;
-	int32 cast_range;
+	float32 path_radius;
+	float32 cast_range;
 	float32 linger_duration;
 	bool m_bTestedGem;
 };
