+++ DumpSource2/schemas/server/CDOTA_Modifier_MonkeyKing_FurArmy_SoldierInPosition.h
@@ -3,7 +3,7 @@ class CDOTA_Modifier_MonkeyKing_FurArmy_SoldierInPosition
 	Vector m_vTargetPos;
 	CHandle< CBaseEntity > m_hAttackTarget;
 	float32 attack_speed;
-	int32 second_radius;
+	float32 second_radius;
 	int32 outer_attack_buffer;
 	CHandle< CBaseEntity > m_hThinker;
 	bool m_bDisarmed;
