+++ DumpSource2/schemas/server/CDOTA_Modifier_Pangolier_Gyroshell.h
@@ -5,8 +5,8 @@ class CDOTA_Modifier_Pangolier_Gyroshell
 	float32 forward_move_speed;
 	float32 turn_rate_boosted;
 	float32 turn_rate;
-	int32 hit_radius;
-	int32 knockback_radius;
+	float32 hit_radius;
+	float32 knockback_radius;
 	int32 damage_pct;
 	bool m_bHitFirstUpdate;
 	GameTime_t m_flHitEndTime;
