+++ DumpSource2/schemas/server/CDOTA_Modifier_Pugna_NetherWard.h
@@ -6,5 +6,5 @@ class CDOTA_Modifier_Pugna_NetherWard
 	int32 attacks_to_destroy;
 	int32 health_restore_pct;
 	int32 mana_restore_pct;
-	int32 self_restoration_range;
+	float32 self_restoration_range;
 };
