+++ DumpSource2/schemas/server/CDOTA_Modifier_Razor_EyeOfTheStorm.h
@@ -1,6 +1,6 @@
 class CDOTA_Modifier_Razor_EyeOfTheStorm
 {
-	int32 radius;
+	float32 radius;
 	int32 armor_reduction;
-	int32 damage;
+	float32 damage;
 };
