+++ DumpSource2/schemas/server/CDOTA_Modifier_TrollWarlord_Whirling_Axes_Melee.h
@@ -1,8 +1,8 @@
 class CDOTA_Modifier_TrollWarlord_Whirling_Axes_Melee
 {
-	int32 damage;
-	int32 hit_radius;
-	int32 axe_movement_speed;
+	float32 damage;
+	float32 hit_radius;
+	float32 axe_movement_speed;
 	float32 whirl_duration;
 	float32 max_range;
 	float32 blind_duration;
